# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API для телефонного справочника",
        "title": "Telephone Book API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/workers": {
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "workers"
                ],
                "summary": "Обновить работника",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Институт",
                        "name": "institute",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Email работника",
                        "name": "email",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "Новые данные работника",
                        "name": "worker",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/workers.UpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/workers.UpdateResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "workers"
                ],
                "summary": "Создать работника",
                "parameters": [
                    {
                        "description": "Данные работника",
                        "name": "worker",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/workers.CreateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/workers.CreateResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "workers"
                ],
                "summary": "Удалить работника",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Институт",
                        "name": "institute",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Email работника",
                        "name": "email",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/workers.DeleteResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/workers/all": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "workers"
                ],
                "summary": "Получить всех работников",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Институт",
                        "name": "institute",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Отдел",
                        "name": "department",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/workers.AllUsersResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/workers/{email}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "workers"
                ],
                "summary": "Получить работника по email",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Email работника",
                        "name": "email",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Институт",
                        "name": "institute",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/workers.GetResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "models.User": {
            "type": "object",
            "properties": {
                "birth_date": {
                    "type": "string"
                },
                "cabinet": {
                    "type": "string"
                },
                "department": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "middle_name": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "string"
                },
                "photo": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "position": {
                    "type": "string"
                },
                "section": {
                    "type": "string"
                },
                "surname": {
                    "type": "string"
                }
            }
        },
...
        "response.Response": {
            "description": "Базовый ответ API",
            "type": "object",
            "properties": {
                "error": {
                    "description": "Сообщение об ошибке, если есть. Пример: \"invalid request\"",
                    "type": "string"
                },
                "status": {
                    "description": "Статус ответа: Ok или Error. Пример: \"Ok\"",
                    "type": "string"
                }
            }
        },
...
        "workers.AllUsersResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "description": "Сообщение об ошибке, если есть. Пример: \"invalid request\"",
                    "type": "string"
                },
                "status": {
                    "description": "Статус ответа: Ok или Error. Пример: \"Ok\"",
                    "type": "string"
                },
                "users": {
                    "description": "Список работников",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.User"
                    }
                }
            }
        },
        "workers.CreateRequest": {
            "type": "object",
            "required": [
                "email",
                "institute",
                "name",
                "phone_number",
                "surname"
            ],
            "properties": {
                "birth_date": {
                    "type": "string"
                },
                "cabinet": {
                    "type": "string"
                },
                "department": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "institute": {
                    "type": "string"
                },
                "middle_name": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "string"
                },
                "position": {
                    "type": "string"
                },
                "section": {
                    "type": "string"
                },
                "surname": {
                    "type": "string"
                }
            }
        },
        "workers.CreateResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "description": "Сообщение об ошибке, если есть. Пример: \"invalid request\"",
                    "type": "string"
                },
                "status": {
                    "description": "Статус ответа: Ok или Error. Пример: \"Ok\"",
                    "type": "string"
                },
                "user_id": {
                    "description": "ID созданного работника",
                    "type": "integer"
                }
            }
        },
        "workers.DeleteResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "description": "Сообщение об ошибке, если есть. Пример: \"invalid request\"",
                    "type": "string"
                },
                "status": {
                    "description": "Статус ответа: Ok или Error. Пример: \"Ok\"",
                    "type": "string"
                }
            }
        },
        "workers.GetResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "description": "Сообщение об ошибке, если есть. Пример: \"invalid request\"",
                    "type": "string"
                },
                "status": {
                    "description": "Статус ответа: Ok или Error. Пример: \"Ok\"",
                    "type": "string"
                },
                "user": {
                    "description": "Работник",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.User"
                        }
                    ]
                }
            }
        },
        "workers.UpdateRequest": {
            "type": "object",
            "required": [
                "cabinet",
                "department",
                "email",
                "institute",
                "name",
                "phone_number",
                "position",
                "surname"
            ],
            "properties": {
                "birth_date": {
                    "type": "string"
                },
                "cabinet": {
                    "type": "string"
                },
                "department": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "institute": {
                    "type": "string"
                },
                "middle_name": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "string"
                },
                "photo": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "position": {
                    "type": "string"
                },
                "section": {
                    "type": "string"
                },
                "surname": {
                    "type": "string"
                }
            }
        },
        "workers.UpdateResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "description": "Сообщение об ошибке, если есть. Пример: \"invalid request\"",
                    "type": "string"
                },
                "status": {
                    "description": "Статус ответа: Ok или Error. Пример: \"Ok\"",
                    "type": "string"
                }
            }
        }
    }
}